requirements.txt
Flask
docx2pdf
python-docx
pdfplumber
